Add Fhir Scope
