+++ data-donation-sdk/src/commonMain/kotlin/care/data4life/datadonation/donation/fhir/anonymization/DateTimeSmearer.kt
@@ -14,7 +14,7 @@
  * contact D4L by email to help@data4life.care.
  */
 
-package care.data4life.datadonation.donation.anonymization
+package care.data4life.datadonation.donation.fhir.anonymization
 
 import care.data4life.datadonation.donation.program.model.BlurFunction
 import care.data4life.hl7.fhir.common.datetime.XsDate
