+++ data-donation-sdk/src/commonTest/kotlin/care/data4life/datadonation/donation/fhir/anonymization/RedactorTest.kt
@@ -14,7 +14,7 @@
  * contact D4L by email to help@data4life.care.
  */
 
-package care.data4life.datadonation.donation.anonymization
+package care.data4life.datadonation.donation.fhir.anonymization
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
